# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Telegram AI Autonomous System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelegramGroup"
    post:
      operationId: createGroup
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramGroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramGroup"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramGroup"
    patch:
      operationId: updateGroup
      tags: [groups]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramGroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramGroup"
    delete:
      operationId: deleteGroup
      tags: [groups]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CHANNELS ──────────────────────────────────────────
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelegramChannel"
    post:
      operationId: createChannel
      tags: [channels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramChannel"
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramChannel"
    patch:
      operationId: updateChannel
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramChannelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramChannel"
    delete:
      operationId: deleteChannel
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── MESSAGES ──────────────────────────────────────────
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: groupId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/start:
    post:
      operationId: startCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/stop:
    post:
      operationId: stopCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ── SCHEDULES ─────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  schemas:
...
    TelegramGroup:
      type: object
      required: [id, telegramId, title, isMonitored, isAutoReply, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        title:
          type: string
        username:
          type: ["string", "null"]
        membersCount:
          type: ["integer", "null"]
        isMonitored:
          type: boolean
        isAutoReply:
          type: boolean
        isAutoModerate:
          type: boolean
        accountId:
          type: ["integer", "null"]
        messagesCount:
          type: integer
        createdAt:
          type: string
    TelegramGroupInput:
      type: object
      required: [telegramId, title]
      properties:
        telegramId:
          type: string
        title:
          type: string
        username:
          type: string
        accountId:
          type: integer
        isMonitored:
          type: boolean
        isAutoReply:
          type: boolean
        isAutoModerate:
          type: boolean
    TelegramGroupUpdate:
      type: object
      properties:
        title:
          type: string
        isMonitored:
          type: boolean
        isAutoReply:
          type: boolean
        isAutoModerate:
          type: boolean
        accountId:
          type: integer

    # ── CHANNEL ──────────────────────────────────────────
    TelegramChannel:
      type: object
      required: [id, telegramId, title, isAutoPost, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        title:
          type: string
        username:
          type: ["string", "null"]
        subscribersCount:
          type: ["integer", "null"]
        isAutoPost:
          type: boolean
        accountId:
          type: ["integer", "null"]
        postsCount:
          type: integer
        createdAt:
          type: string
    TelegramChannelInput:
      type: object
      required: [telegramId, title]
      properties:
        telegramId:
          type: string
        title:
          type: string
        username:
          type: string
        accountId:
          type: integer
        isAutoPost:
          type: boolean
    TelegramChannelUpdate:
      type: object
      properties:
        title:
          type: string
        isAutoPost:
          type: boolean
        accountId:
          type: integer

    # ── MESSAGE ──────────────────────────────────────────
    Message:
      type: object
      required: [id, content, direction, status, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: ["integer", "null"]
        groupId:
          type: ["integer", "null"]
        channelId:
          type: ["integer", "null"]
        content:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
          enum: [sent, delivered, failed, pending]
        isAIGenerated:
          type: boolean
        senderUsername:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        accountId:
          type: integer
        groupId:
          type: integer
        channelId:
          type: integer
        content:
          type: string
        useAI:
          type: boolean
        personalityId:
          type: integer

    # ── CAMPAIGN ─────────────────────────────────────────
    Campaign:
      type: object
      required: [id, name, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [motivation, crypto, marketing, announcement, support, custom]
        status:
          type: string
          enum: [draft, active, paused, completed]
        personalityId:
          type: ["integer", "null"]
        targetGroups:
          type: array
          items:
            type: integer
        targetChannels:
          type: array
          items:
            type: integer
        messagesSent:
          type: integer
        intervalHours:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [motivation, crypto, marketing, announcement, support, custom]
        personalityId:
          type: integer
        targetGroups:
          type: array
          items:
            type: integer
        targetChannels:
          type: array
          items:
            type: integer
        intervalHours:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        personalityId:
          type: integer
        intervalHours:
          type: integer
        startDate:
          type: string
        endDate:
          type: string

    # ── SCHEDULE ─────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [info, warning, error, success, security, ai]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        accountId:
          type: ["integer", "null"]
        createdAt:
          type: string