# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram AI Autonomous System API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  accountId?: number;
  groupId?: number;
  channelId?: number;
  content: string;
  useAI?: boolean;
  personalityId?: number;
}