# lib/api-zod/src/generated/types/telegramGroupUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram AI Autonomous System API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramGroupUpdate {
  title?: string;
  isMonitored?: boolean;
  isAutoReply?: boolean;
  isAutoModerate?: boolean;
  accountId?: number;
}